Refactor command subscriptions
